refactor(MemberService): separate authentication logic into AuthService and simplify email conflict check logic
