+++ src/main/java/gift/repository/MemberRepository.java
@@ -1,13 +1,14 @@
 package gift.repository;
 
 import gift.entity.Member;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByEmail(String email);
+
+    boolean existsByEmail(String email);
 }
