+++ src/main/java/gift/service/MemberService.java
@@ -1,55 +1,44 @@
 package gift.service;
 
-import gift.dto.AuthenticatedMember;
-import gift.dto.LoginMemberCommand;
 import gift.dto.MemberDto;
 import gift.dto.NewMemberCommand;
 import gift.dto.UpdateMemberCommand;
 import gift.entity.Member;
 import gift.exception.ConflictException;
-import gift.exception.InvalidCredentialsException;
 import gift.exception.NotFoundException;
 import gift.repository.MemberRepository;
-import gift.token.JwtTokenProvider;
 import gift.util.BCryptEncryptor;
 import gift.util.PasswordUtility;
-import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    private final JwtTokenProvider jwtTokenProvider;
 
-    public MemberService(MemberRepository memberRepository, JwtTokenProvider jwtTokenProvider) {
+    public MemberService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
-        this.jwtTokenProvider = jwtTokenProvider;
     }
 
+    // Create
     @Transactional
     public MemberDto createMember(NewMemberCommand newMemberCommand) {
-        checkValidMemberUpdate(newMemberCommand.email(), null);
-
+        if (memberRepository.existsByEmail(newMemberCommand.email())) {
+            throw new ConflictException("Email already in use: email=" + newMemberCommand.email());
+        }
         String encodedPassword = BCryptEncryptor.encrypt(newMemberCommand.password());
         Member member = new Member(newMemberCommand.email(), encodedPassword,
                 newMemberCommand.role());
         return MemberDto.from(memberRepository.save(member));
     }
 
-    @Transactional(readOnly = true)
-    public String login(LoginMemberCommand command) {
-        Optional<Member> optionalMember = memberRepository.findByEmail(command.email());
-        if (optionalMember.isEmpty() || !BCryptEncryptor.matches(command.password(),
-                optionalMember.get().getPassword())) {
-            throw new InvalidCredentialsException("이메일 또는 비밀번호가 올바르지 않습니다.");
-        }
-        return jwtTokenProvider.createToken(optionalMember.get());
+    // Read
+    protected Member findMemberById(Long id) {
+        return memberRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Member not found: id=" + id));
     }
 
     @Transactional(readOnly = true)
@@ -64,10 +53,15 @@ public class MemberService {
         return MemberDto.from(member);
     }
 
+    // Update
     @Transactional
     public MemberDto updateMember(UpdateMemberCommand updateMemberCommand) {
         Member member = findMemberById(updateMemberCommand.id());
-        checkValidMemberUpdate(updateMemberCommand.email(), member.getId());
+        if (!member.getEmail().equals(updateMemberCommand.email())
+                && memberRepository.existsByEmail(updateMemberCommand.email())) {
+            throw new ConflictException(
+                    "Email already in use: email=" + updateMemberCommand.email());
+        }
         String rawPassword = null;
         String encodedPassword = null;
         if (updateMemberCommand.resetPassword()) {
@@ -78,42 +72,12 @@ public class MemberService {
         return MemberDto.from(member, rawPassword);
     }
 
+    // Delete
     @Transactional
     public void deleteMember(Long id) {
         if (!memberRepository.existsById(id)) {
             throw new NotFoundException("Member not found: id=" + id);
         }
         memberRepository.deleteById(id);
-    }
-
-    @Transactional(readOnly = true)
-    public AuthenticatedMember getAuthenticationFromToken(String token) {
-        if (token == null || !jwtTokenProvider.validateToken(token)) {
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid token");
-        }
-        String username = jwtTokenProvider.getUsername(token);
-        Optional<Member> optionalMember = memberRepository.findByEmail(username);
-        if (optionalMember.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid token");
-        }
-        return AuthenticatedMember.from(optionalMember.get());
-    }
-
-    protected Member findMemberById(Long id) {
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Member not found: id=" + id));
-    }
-
-    private void checkValidMemberUpdate(String email, Long memberId) {
-        if (!isEmailUsable(email, memberId)) {
-            throw new ConflictException("Email already in use: email=" + email);
-        }
-    }
-
-    // 호출 메소드의 Transaction에 참여
-    private boolean isEmailUsable(String email, Long memberId) {
-        Optional<Member> optionalMember = memberRepository.findByEmail(email);
-        // 해당 이메일을 사용중인 멤버가 없거나, 이를 요청한 회원이 해당 이메일의 소유자일 경우 (즉, 이메일 변경이 아님)
-        return optionalMember.isEmpty() || optionalMember.get().getId().equals(memberId);
     }
 }
