+++ src/test/java/gift/service/MemberServiceTest.java
@@ -8,16 +8,13 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 
-import gift.dto.LoginMemberCommand;
 import gift.dto.MemberDto;
 import gift.dto.NewMemberCommand;
 import gift.dto.UpdateMemberCommand;
 import gift.entity.Member;
 import gift.entity.Role;
 import gift.exception.ConflictException;
-import gift.exception.InvalidCredentialsException;
 import gift.repository.MemberRepository;
-import gift.token.JwtTokenProvider;
 import gift.util.BCryptEncryptor;
 import gift.util.PasswordUtility;
 import java.util.Optional;
@@ -29,19 +26,12 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.web.server.ResponseStatusException;
 
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
 
     @Mock
-    private BCryptEncryptor bCryptEncryptor;
-
-    @Mock
     private MemberRepository memberRepository;
-
-    @Mock
-    private JwtTokenProvider jwtTokenProvider;
 
     @InjectMocks
     private MemberService memberService;
@@ -59,7 +49,7 @@ class MemberServiceTest {
 
             Member expectedMember = new Member(1L, email, hashedPassword, Role.ROLE_USER);
 
-            when(memberRepository.findByEmail(email)).thenReturn(Optional.empty());
+            when(memberRepository.existsByEmail(email)).thenReturn(false);
 
             try (MockedStatic<BCryptEncryptor> encryptor = mockStatic(BCryptEncryptor.class)) {
 
@@ -85,72 +75,10 @@ class MemberServiceTest {
             String email = "existing@email.com";
             String rawPassword = "password123456789";
 
-            when(memberRepository.findByEmail(email)).thenReturn(
-                    Optional.of(new Member(100L, email, "hashedPassword", null)));
+            when(memberRepository.existsByEmail(email)).thenReturn(true);
 
             assertThrows(ConflictException.class,
                     () -> memberService.createMember(new NewMemberCommand(email, rawPassword)));
-        }
-    }
-
-    @Nested
-    @DisplayName("Member login() - 로그인 테스트")
-    class LoginTests {
-
-        @Test
-        @DisplayName("정상적인 로그인 시 토큰 반환")
-        void 정상적인_로그인_시_토큰반환() {
-            String email = "registered@email.com";
-            String rawPassword = "password123456789";
-            String hashedPassword = "hashedPassword";
-            Member existingMember = new Member(1L, email, hashedPassword, Role.ROLE_USER);
-
-            when(memberRepository.findByEmail(email)).thenReturn(Optional.of(existingMember));
-
-            try (MockedStatic<BCryptEncryptor> encryptor = mockStatic(BCryptEncryptor.class)) {
-
-                encryptor.when(() -> BCryptEncryptor.matches(rawPassword, hashedPassword))
-                        .thenReturn(true);
-                when(jwtTokenProvider.createToken(existingMember)).thenReturn("validToken");
-
-                String token = memberService.login(new LoginMemberCommand(email, rawPassword));
-
-                assertThat(token).isEqualTo("validToken");
-            }
-        }
-
-        @Test
-        @DisplayName("존재하지 않는 이메일로 로그인 시 예외 발생")
-        void 존재하지_않는_이메일로_로그인_시_예외발생() {
-            String email = "not.registered@email.com";
-            String rawPassword = "password123456789";
-            String hashedPassword = "hashedPassword";
-
-            when(memberRepository.findByEmail(email)).thenReturn(Optional.empty());
-
-            assertThrows(InvalidCredentialsException.class,
-                    () -> memberService.login(new LoginMemberCommand(email, rawPassword)));
-        }
-
-        @Test
-        @DisplayName("잘못된 비밀번호로 로그인 시 예외 발생")
-        void 잘못된_비밀번호로_로그인_시_예외발생() {
-            String email = "registered@email.com";
-            String rawPassword = "wrongPassword123456789";
-            String hashedPassword = "hashedPassword";
-
-            Member existingMember = new Member(1L, email, hashedPassword, Role.ROLE_USER);
-
-            when(memberRepository.findByEmail(email)).thenReturn(Optional.of(existingMember));
-
-            try (MockedStatic<BCryptEncryptor> encryptor = mockStatic(BCryptEncryptor.class)) {
-
-                encryptor.when(() -> BCryptEncryptor.matches(rawPassword, hashedPassword))
-                        .thenReturn(false);
-
-                assertThrows(InvalidCredentialsException.class,
-                        () -> memberService.login(new LoginMemberCommand(email, rawPassword)));
-            }
         }
     }
 
@@ -170,7 +98,6 @@ class MemberServiceTest {
 
             Member existingMember = new Member(id, email, hashedPassword, role);
             when(memberRepository.findById(id)).thenReturn(Optional.of(existingMember));
-            when(memberRepository.findByEmail(newEmail)).thenReturn(Optional.empty());
 
             MemberDto result = memberService.updateMember(
                     new UpdateMemberCommand(id, newEmail, false, newRole));
@@ -196,7 +123,6 @@ class MemberServiceTest {
 
             Member existingMember = new Member(id, email, hashedPassword, role);
             when(memberRepository.findById(id)).thenReturn(Optional.of(existingMember));
-            when(memberRepository.findByEmail(newEmail)).thenReturn(Optional.empty());
 
             try (
                     MockedStatic<BCryptEncryptor> encryptor = mockStatic(BCryptEncryptor.class);
@@ -226,41 +152,4 @@ class MemberServiceTest {
 
     // 단순 CRUD 테스트 생략
 
-    @Nested
-    @DisplayName("Member getAuthenticationFromToken() - 토큰으로 인증 정보 조회 테스트")
-    class GetAuthenticationFromTokenTests {
-
-        @Test
-        @DisplayName("유효한 토큰으로 인증 정보 조회")
-        void 유효한_토큰으로_인증정보조회() {
-            String username = "email@email.com";
-            String token = "validToken";
-            Member member = new Member(1L, "email", "hashedPassword", Role.ROLE_USER);
-            when(jwtTokenProvider.validateToken(token)).thenReturn(true);
-            when(jwtTokenProvider.getUsername(token)).thenReturn(username);
-            when(memberRepository.findByEmail(username)).thenReturn(Optional.of(member));
-            assertThat(memberService.getAuthenticationFromToken(token)).isNotNull();
-        }
-
-        @Test
-        @DisplayName("유효하지 않은 토큰으로 인증 정보 조회 시 예외 발생")
-        void 유효하지_않은_토큰으로_인증정보조회시_예외발생() {
-            String token = "validToken";
-            when(jwtTokenProvider.validateToken(token)).thenReturn(false);
-            assertThrows(ResponseStatusException.class,
-                    () -> memberService.getAuthenticationFromToken(token));
-        }
-
-        @Test
-        @DisplayName("유효한 토큰이지만 회원 정보가 없는 경우 예외 발생")
-        void 유효한_토큰이지만_회원정보가_없는_경우_예외발생() {
-            String username = "deletedMember@email.com";
-            String token = "validToken";
-            when(jwtTokenProvider.validateToken(token)).thenReturn(true);
-            when(jwtTokenProvider.getUsername(token)).thenReturn(username);
-            when(memberRepository.findByEmail(username)).thenReturn(Optional.empty());
-            assertThrows(ResponseStatusException.class,
-                    () -> memberService.getAuthenticationFromToken(token));
-        }
-    }
 }
